+++ docs/CHANGELOG.md
@@ -30,6 +30,7 @@ sections to include in release notes:
 - Changelog (#33)
 - Build docs and push to gh-pages (#41)
 - events.out file for agronomic event handling (#57)
+- utility `tools/trim_first_chars.sh` to trim the first n characters from every row in a file, useful for updating old input files to remove location column
 
 ### Fixed
 
@@ -40,12 +41,16 @@ sections to include in release notes:
 
 - Reorganized codebase (#34, #37)
 - Deprecated: "RUNTYPE" is obsolete. Will be silently ignored if set to 'standard' or error if set to anything else. Runs in 'standard' mode by default.
+- Deprecated: "LOCATION" is obsolete. Will be ignored with warning. (#92)
+- Deprecated: All columns in *.param except for name and value. Will be ignored with warning. (#92)
+- Deprecated: location column in input climate files. Will be ignored with warning. (#92)
 
 ### Removed
 
 - Removed many experimental sites, data, and executable code as part of reorg (#34, #37)
 - Removed obsolete run types senstest and montecarlo and associated code (#69, #76)
 - Removed obsolete estimate program and associated code (#70, #82)
+- Removed multi-site support; in particular, output files no longer have a location column (#92)
 
 ### Git SHA
 [TBD]
@@ -148,5 +153,3 @@ Braswell, Bobby H., William J. Sacks, Ernst Linder, and David S. Schimel. 2005. 
 Sacks, William J., David S. Schimel, and Russell K. Monson. 2007. “Coupling between Carbon Cycling and Climate in a High-Elevation, Subalpine Forest: A Model-Data Fusion Analysis.” Oecologia 151 (1): 54–68. https://doi.org/10.1007/s00442-006-0565-2.
 
 Sacks, William J., David S. Schimel, Russell K. Monson, and Bobby H. Braswell. 2006. “Model‐data Synthesis of Diurnal and Seasonal CO2 Fluxes at Niwot Ridge, Colorado.” Global Change Biology 12 (2): 240–59. https://doi.org/10.1111/j.1365-2486.2005.01059.x.
-
-